+++ dev/FormsPlotSandbox/FormExperimentalTick.Designer.cs
@@ -1,0 +1,99 @@
+﻿namespace FormsPlotSandbox
+{
+    partial class FormExperimentalTick
+    {
+        /// <summary>
+        /// Required designer variable.
+        /// </summary>
+        private System.ComponentModel.IContainer components = null;
+
+        /// <summary>
+        /// Clean up any resources being used.
+        /// </summary>
+        /// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
+        protected override void Dispose(bool disposing)
+        {
+            if (disposing && (components != null))
+            {
+                components.Dispose();
+            }
+            base.Dispose(disposing);
+        }
+
+        #region Windows Form Designer generated code
+
+        /// <summary>
+        /// Required method for Designer support - do not modify
+        /// the contents of this method with the code editor.
+        /// </summary>
+        private void InitializeComponent()
+        {
+            this.components = new System.ComponentModel.Container();
+            this.nudLineThickness = new System.Windows.Forms.NumericUpDown();
+            this.label1 = new System.Windows.Forms.Label();
+            this.formsPlot1 = new ScottPlot.FormsPlot();
+            ((System.ComponentModel.ISupportInitialize)(this.nudLineThickness)).BeginInit();
+            this.SuspendLayout();
+            // 
+            // nudLineThickness
+            // 
+            this.nudLineThickness.Location = new System.Drawing.Point(12, 26);
+            this.nudLineThickness.Maximum = new decimal(new int[] {
+            10,
+            0,
+            0,
+            0});
+            this.nudLineThickness.Name = "nudLineThickness";
+            this.nudLineThickness.Size = new System.Drawing.Size(68, 20);
+            this.nudLineThickness.TabIndex = 1;
+            this.nudLineThickness.Value = new decimal(new int[] {
+            2,
+            0,
+            0,
+            0});
+            this.nudLineThickness.ValueChanged += new System.EventHandler(this.nudLineThickness_ValueChanged);
+            // 
+            // label1
+            // 
+            this.label1.AutoSize = true;
+            this.label1.Location = new System.Drawing.Point(9, 10);
+            this.label1.Name = "label1";
+            this.label1.Size = new System.Drawing.Size(71, 13);
+            this.label1.TabIndex = 2;
+            this.label1.Text = "line thickness";
+            // 
+            // formsPlot1
+            // 
+            this.formsPlot1.Anchor = ((System.Windows.Forms.AnchorStyles)((((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Bottom) 
+            | System.Windows.Forms.AnchorStyles.Left) 
+            | System.Windows.Forms.AnchorStyles.Right)));
+            this.formsPlot1.Location = new System.Drawing.Point(12, 52);
+            this.formsPlot1.Name = "formsPlot1";
+            this.formsPlot1.Size = new System.Drawing.Size(776, 386);
+            this.formsPlot1.TabIndex = 0;
+            // 
+            // FormExperimentalTick
+            // 
+            this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
+            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            this.ClientSize = new System.Drawing.Size(800, 450);
+            this.Controls.Add(this.label1);
+            this.Controls.Add(this.nudLineThickness);
+            this.Controls.Add(this.formsPlot1);
+            this.Name = "FormExperimentalTick";
+            this.StartPosition = System.Windows.Forms.FormStartPosition.CenterScreen;
+            this.Text = "Experimental Tick Tester";
+            this.Load += new System.EventHandler(this.FormExperimentalTick_Load);
+            ((System.ComponentModel.ISupportInitialize)(this.nudLineThickness)).EndInit();
+            this.ResumeLayout(false);
+            this.PerformLayout();
+
+        }
+
+        #endregion
+
+        private ScottPlot.FormsPlot formsPlot1;
+        private System.Windows.Forms.NumericUpDown nudLineThickness;
+        private System.Windows.Forms.Label label1;
+    }
+}